bdev/qos: cleanup the IOs on QoS thread when removing IO channel
The I/O channel can be obtained and released dynamically on
demand, in the case of QoS has been enabled. The I/Os could
be queued on the QoS thread due to rate limiting. Considering
the thread safe operation, the thread for the I/O channel and
QoS thread is mostly different. The asynchronous method of
message passing through threads is used to drain the I/Os when
"spdk_put_io_channel()" is executed to cleanup the I/O channel.

In this patch:
1. Add a new member async_destroying for the spdk_io_channel
structure so that spdk_put_io_channel() API can suppport the
asynchronous I/Os draining on QoS thread.
2. UT code added to release I/O channel at runtime.

+++ lib/util/io_channel.c
@@ -448,6 +448,7 @@ spdk_get_io_channel(void *io_device)
 	ch->destroy_cb = dev->destroy_cb;
 	ch->thread = thread;
 	ch->ref = 1;
+	ch->async_destroying = false;
 	TAILQ_INSERT_TAIL(&thread->io_channels, ch, tailq);
 
 	dev->refcnt++;
@@ -476,7 +477,12 @@ _spdk_put_io_channel(void *arg)
 	assert(ch->thread == spdk_get_thread());
 	assert(ch->ref == 0);
 
-	ch->destroy_cb(ch->dev->io_device, spdk_io_channel_get_ctx(ch));
+	if (ch->async_destroying == false) {
+		ch->destroy_cb(ch->dev->io_device, spdk_io_channel_get_ctx(ch));
+		if (ch->async_destroying == true) {
+			return;
+		}
+	}
 
 	pthread_mutex_lock(&g_devlist_mutex);
 	ch->dev->refcnt--;
@@ -500,6 +506,11 @@ _spdk_put_io_channel(void *arg)
 void
 spdk_put_io_channel(struct spdk_io_channel *ch)
 {
+	if (ch->async_destroying == true) {
+		spdk_thread_send_msg(ch->thread, _spdk_put_io_channel, ch);
+		return;
+	}
+
 	ch->ref--;
 
 	if (ch->ref == 0) {
